+++ src/main/java/io/github/jperparas/resourcetrackerpwa/repositories/PcaRepository.java
@@ -3,5 +3,5 @@ package io.github.jperparas.resourcetrackerpwa.repositories;
 import io.github.jperparas.resourcetrackerpwa.entities.Pca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PcaRepository extends JpaRepository<Pca, Long> {
+public interface PcaRepository extends JpaRepository<Pca, Integer> {
 }
