+++ src/main/java/io/github/jperparas/resourcetrackerpwa/repositories/SpotRepository.java
@@ -3,5 +3,5 @@ package io.github.jperparas.resourcetrackerpwa.repositories;
 import io.github.jperparas.resourcetrackerpwa.entities.Spot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpotRepository extends JpaRepository<Spot, Long> {
+public interface SpotRepository extends JpaRepository<Spot, Integer> {
 }
